feat(as/policy): split opa into fs/in-memory engines for fs-free eval
Decompose the monolithic opa/mod.rs into:
- opa/mod.rs: shared helpers (is_valid_policy_id, policy_uses_legacy_reference,
  reference-value lookup) used by both backends;
- opa/fs.rs: the existing filesystem-backed OPA, gated behind the `fs` feature;
- opa/in_memory.rs: a new OPAInMemory engine that holds rego policy sources
  in a RwLock<HashMap> and evaluates them with Regorus, mirroring OPAs
  engine usage without any filesystem access.

Gate the fs-based OPA struct and PolicyEngineType::OPA/to_policy_engine behind
the `fs` feature so the policy layer can be compiled and exercised without a
filesystem. OPAInMemory::with_raw_default_policy preloads the default policy
under its id stem (.rego stripped) so broker default-policy evaluation
(evaluate(..., "default", ...)) succeeds without disk access.
